HUB de Vida Añadido
Se añadieron dos HUB de vida, uno siendo una barra de vida que esta oculto y otro la label escrito

+++ Desafios_M_Gundic/Assets/Script/Curar.cs
@@ -6,7 +6,7 @@ public class Curar : MonoBehaviour
 {
     // Variables a configurar desde el editor
     [Header("Configuracion")]
-    [SerializeField] float puntos = 1f;
+    [SerializeField] int puntos = 1;
 
     private void OnParticleCollision(GameObject other)
     {
